Actionneur/Capteur correction

+++ Actionneur.java
@@ -1,17 +1,17 @@
 public class Actionneur extends Object{
 
-    private int id ;
+    private long id ;
     private String Etats ;
     private int valeur ;
 
-    public Actionneur(int id,String Etats,int valeur) {
+    public Actionneur(long id,String Etats,int valeur) {
         this.id = id ;
         this.Etats = Etats;
         this.valeur = valeur;
     }
 
     public void faire() {
-        
+
     }
